requirements.txt
google-ads==13.0.0
google-cloud-bigquery==2.26.0
pandas==1.3.4
pyarrow==6.0.1
